refactor legacy, rm sidebar
