+++ src/routes/Legacy/hooks/useHighScores.ts
@@ -1,4 +1,5 @@
-import { legacySlice, useAppDispatch, useAppSelector } from "../../../store";
+import { useAppDispatch, useAppSelector } from "../../../store";
+import { legacySlice } from "../../../store/legacy";
 
 const { actions } = legacySlice;
