+++ src/store/legacy/index.ts
@@ -1,6 +1,5 @@
 import { createSlice } from "@reduxjs/toolkit";
 import type { PayloadAction } from "@reduxjs/toolkit";
-
 import { high_scores } from "./high_scores";
 
 interface LegacyState {
